requirements.txt
tqdm~=4.66.1
beautifulsoup4~=4.12.2
lxml~=5.1.0
requests~=2.31.0
pillow~=10.2.0
aiodecorators~=0.2.1
numpy~=1.26.3
py360convert~=0.1.0
natsort~=8.4.0